🏷️ Molecule > Input: type 병합, 위치 이동
