--- client/src/components/molecules/Input/Label/type.ts
@@ -1,13 +1,0 @@
-import { InputProps } from "atoms/Input/type";
-
-export interface LabelInputProps extends InputProps {
-  label: string;
-  $input: HTMLInputElement;
-  className?: string;
-  hasBoundary?: boolean;
-}
-
-export interface CheckBoxInputProps {
-  id?: string;
-  onClick?: (e: ClickEvent<HTMLInputElement>) => void;
-}
